+++ src/vec/vec2.rs
@@ -127,9 +127,9 @@ macro_rules! vec2s {
 
             #[inline]
             pub fn normalize(&mut self) {
-                let mag = self.mag();
-                self.x /= mag;
-                self.y /= mag;
+                let r_mag = $t::splat(1.0) /self.mag();
+                self.x *= mag;
+                self.y *= mag;
             }
 
             #[inline]