+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 - Add Mat3 into Rotor3 conversion for rotation matrices
+- Remove heavy reliance on `mul_add` due to negligible performance benefit and in many cases performance detriment.
+- Slightly optimize Vector `normalize`.
 
 ## 0.6.1
 - Add scalar multiplication and componentwise addition for `MatN`