+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove heavy reliance on `mul_add` due to negligible performance benefit and in many cases performance detriment.
 - Slightly optimize Vector `normalize`.
 - `Rotor2::from_angle_plane()` now takes plane and angle as separate arguments.
+- Add `MatN::adjugate()`
 
 ## 0.6.1
 - Add scalar multiplication and componentwise addition for `MatN`