+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 <!-- next-header -->
-##Unreleased
+## Unreleased
+
 ## 0.7.0
 - Add Mat3 into Rotor3 conversion for rotation matrices
 - Remove heavy reliance on `mul_add` due to negligible performance benefit and in many cases performance detriment.